// src/types/router.ts
import type { Locale } from './i18n'
import type { Color } from '@/types/color'

export enum RouteName {
  // en routes
  HOME_EN = 'HOME_EN',
  PROJECTS_EN = 'PROJECTS_EN',
  BOOKING_ENGINE_EN = 'BOOKING_ENGINE_EN',
  CLINICAL_TRIALS_EN = 'CLINICAL_TRIALS_EN',
  CORONAVIRUS_VISUALIZATION_EN = 'CORONAVIRUS_VISUALIZATION_EN',
  SCIENTIFIC_PUBLICATIONS_EN = 'SCIENTIFIC_PUBLICATIONS_EN',
  STATISTICAL_ANALYSIS_EN = 'STATISTICAL_ANALYSIS_EN',
  TICKETS_SALE_OPTIMIZATION_EN = 'TICKETS_SALE_OPTIMIZATION_EN',
  EXHIBITION_MANAGEMENT_EN = 'EXHIBITION_MANAGEMENT_EN',
  VEGANSURVIVE_COMMUNITY_EN = 'VEGANSURVIVE_COMMUNITY_EN',
  TOURISM_TRAFFIC_ARBITRAGE_EN = 'TOURISM_TRAFFIC_ARBITRAGE_EN',
  PAID_TRAFFIC_EN = 'PAID_TRAFFIC',
  WEB_SCRAPING_EN = 'WEB_SCRAPING_EN',
  PERSONAL_PORTFOLIO_EN = 'PERSONAL_PORTFOLIO_EN',
  FLUID_FINANCE_EN = 'FLUID_FINANCE_EN',
  DEEP_DAO_EN = 'DEEP_DAO_EN',
  CONTACTS_EN = 'CONTACTS_EN',

  // ru routes
  HOME_RU = 'HOME_RU',
  ROOT_RU = 'ROOT_RU',
  PROJECTS_RU = 'PROJECTS_RU',
  BOOKING_ENGINE_RU = 'BOOKING_ENGINE_RU',
  CLINICAL_TRIALS_RU = 'CLINICAL_TRIALS_RU',
  CORONAVIRUS_VISUALIZATION_RU = 'CORONAVIRUS_VISUALIZATION_RU',
  SCIENTIFIC_PUBLICATIONS_RU = 'SCIENTIFIC_PUBLICATIONS_RU',
  STATISTICAL_ANALYSIS_RU = 'STATISTICAL_ANALYSIS_RU',
  TICKETS_SALE_OPTIMIZATION_RU = 'TICKETS_SALE_OPTIMIZATION_RU',
  EXHIBITION_MANAGEMENT_RU = 'EXHIBITION_MANAGEMENT_RU',
  TOURISM_TRAFFIC_ARBITRAGE_RU = 'TOURISM_TRAFFIC_ARBITRAGE_RU',
  VEGANSURVIVE_COMMUNITY_RU = 'VEGANSURVIVE_COMMUNITY_RU',
  PAID_TRAFFIC_RU = 'PAID_TRAFFIC_RU',
  WEB_SCRAPING_RU = 'WEB_SCRAPING_RU',
  PERSONAL_PORTFOLIO_RU = 'PERSONAL_PORTFOLIO_RU',
  FLUID_FINANCE_RU = 'FLUID_FINANCE_RU',
  DEEP_DAO_RU = 'DEEP_DAO_RU',
  CONTACTS_RU = 'CONTACTS_RU',

  NOT_FOUND_EN = 'NOT_FOUND_EN',
  NOT_FOUND_RU = 'NOT_FOUND_RU'
}

export interface Theme {
  color: Color;
  appleTouchIcon: () => Promise<typeof import('*.png')>;
  appleTouchIcon180: () => Promise<typeof import('*.png')>;
  favicon32: () => Promise<typeof import('*.png')>;
}

export enum SearchBot {
  INDEX_FOLLOW = 'index,follow',
  NOINDEX_NOFOLLOW = 'noindex,nofollow'
}

// Routes to the current page on remained languages.
// E.g. page in EN is viewed => links to RU/UK pages should be presented.
// Page is in UK => links to EN/RU pages should be presented.
export type ComplementI18nRoutes = Partial<Record<Locale, { name: RouteName; }>>