// src/types/technology.ts
export enum Technology {
  VANILLA_JS = 'Vanilla JS',
  VUE_JS = 'Vue.js',
  VUE_3 = 'Vue 3',
  PINIA = 'Pinia',
  VUEX = 'Vuex',
  AXIOS = 'Axios',
  DECIMAL_JS = 'decimal.js',
  JWT = 'JWT',
  VUE_ROUTER = 'Vue Router',
  VITE = 'vite',
  ANGULAR_JS = 'AngularJS',
  REACT = 'React',
  GATSBY = 'Gatsby',
  DOT_NET= '.NET',
  ASP_NET = 'ASP.NET',
  PHP = 'php',
  UNIT_TEST = 'unit-test',

  WEB3 = 'Web3',
  WALLET_CONNECT = 'Web3Modal (Wallet Connect)',
  ETHERS_JS = 'ethers.js',
  WAGMI = 'wagmi',

  SCSS = 'scss',
  LESS = 'less',
  MATERIALIZECSS = 'materializecss',
  BOOTSTRAP = 'bootstrap',

  APEX_CHARTS = 'ApexCharts',
  VUE_I18N = 'Vue I18n',
  VUELIDATE = 'Vuelidate',
  LODASH = 'lodash',
  DATE_FNS = 'date-fns',

  CHEERIO = 'cheerio',
  NODE_JS = 'Node.js',
  PASSPORT_OAUTH = 'passport/OAuth',
  EXPRESS = 'Express',

  MONGO_DB = 'MongoDB',
  MS_SQL = 'MS SQL',
  ORACLE = 'Oracle',
  MY_SQL = 'MySQL',

  SOFT_SKILLS = 'soft skills',
  SEO_SMM = 'SEO/SMM',
  SSG = 'SSG',
  SEO = 'SEO',

  WPML = 'WPML',
  WORDPRESS = 'Wordpress',
  WOO_COMMERCE = 'WooCommerce'
}