// src/types/tool.ts
export enum Tool {
  JIRA = 'JIRA',
  BITBUCKET = 'Bitbucket',
  GIT = 'git',
  SVN = 'svn',
  FIGMA = 'Figma',
  PHOTOSHOP = 'Photoshop',
  GOOGLE_ADS = 'Google Ads',
  YANDEX_METRIKA = 'Yandex.Metrika',
  YANDEX_DIRECT = 'Yandex.Direct',
  YANDEX_WORDSTAT = 'Yandex.Wordstat',
  BLENDER = 'Blender',
  MAX_3DS = '3ds Max',
  ADOBE_PREMIERE_PRO = 'Adobe Premiere Pro',
  FL_STUDIO = 'FL Studio',
  UI_UX = 'UI/UX'
}